requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Werkzeug==2.3.7
email-validator==2.0.0
qrcode==7.4.2
reportlab==4.0.4
Pillow==10.0.0
python-dotenv==1.0.0
Flask-WTF==1.1.1
pytest==7.4.0
itsdangerous==2.1.2
requests==2.31.0
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
eventlet==0.33.3
gunicorn==21.2.0
